requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.2
numpy>=1.24.0
PyPDF2>=3.0.1
python-docx>=0.8.11
chardet>=5.2.0
requests>=2.31.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
